requirements.txt
fastapi[all]
langchain
google-search-results
openai
requests
python-dotenv
supabase
playwright
gpt-engineer
deeplake